requirements.txt
python-dotenv==0.19.2
webbot==0.34
selenium==4.2.0
beautifulsoup4==4.10.0
localStoragePy==0.2.3